+++ lib/fleck/client/request.rb
@@ -42,7 +42,7 @@ module Fleck
       }, mode: :compat)
       logger.debug("Sending request with data: #{data}")
 
-      @exchange.publish(data, routing_key: @routing_key, reply_to: @reply_to, correlation_id: @id)
+      @exchange.publish(data, routing_key: @routing_key, reply_to: @reply_to, correlation_id: @id, mandatory: true)
       @lock.synchronize { @condition.wait(@lock) } unless async
     end